feat(后台管理): 栏目管理
完善菜单

+++ waynboot-common/src/main/java/com/wayn/common/core/domain/shop/ColumnGoodsRelation.java
@@ -29,12 +29,12 @@ public class ColumnGoodsRelation implements Serializable {
     /**
      * 栏目ID
      */
-    private Integer columnId;
+    private Long columnId;
 
     /**
      * 商品ID
      */
-    private Integer goodsId;
+    private Long goodsId;
 
 
 }
